revert Allow no applications in the navring
For those of us that do not have Google installed, or otherwise not
wanting the navring, this patch enables "None" to be chosen for the
center of the navring. This was disallowed in lollipop for some
reason.

This does require changing a previous decision to always show the middle
target. I'm not sure if that's done just as a design decision, a Google
CTS requirement, or just simplicity. The previous commit[1] and gerrit
changeset[2] does not reveal this point as far as I can tell.

Change-Id: I992266eb6016e61f0ff01270386eed565dd495e5 (reverted from commit ec3131fc1b8e70e0fe40090bef85a6bb13382dc4)