Squashed On-The-Go mode (1/2)
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Aicpfy:
- Use Aicp utils
- Proper icon color from Aicp LP

Change-Id: Ibbe1e288c53cd7f53196f0a9110807d907738408

Conflicts:
	core/java/com/android/internal/util/cmremix/CMRemixUtils.java
	core/res/res/values/aicp_strings.xml
	core/res/res/values/aicp_symbols.xml
	packages/SystemUI/res/values/aicp_config.xml
	packages/SystemUI/res/values/aicp_strings.xml