revert: telephony: Hack GSM and LTE signal strength
makeSignalStrengthFromRilParcel() is used for both solicited and
unsolicited requests, so we can hack it to send better signal strength
values to the modem backend. This significantly helps in areas where
GSM or LTE signal is poor but usable, in which case the modem would often
disconnect completely or fall back to a lower network mode (resulting in
poor UX).

Revised to fix the invalid GSM signal strength values (over 31 ASU).

Change-Id: Id2daa40b96bfe226faa2cae3c8e349fb3201ad13 (reverted from commit 7a74cc0aec2ba8f0a29914ca3df6c25b6f03e6da)