SystemUI: stop hiding hotspot & color inversion tiles
These were hidden based on usage patterns to not clutter up the initial
dropdown with a bunch of tiles the user may not user.

Since we allow configuration of the tiles, the tile pages are
limited in height, and it doesn't really fit in with our edit tile
scheme, let's always show these tiles. The user can hide them if they
need.

This provides for a more consistent experience when hitting "edit
tiles."

Change-Id: If4c1254d2200f260a1263658eab6761755420816
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/qs/tiles/ColorInversionTile.java
	packages/SystemUI/src/com/android/systemui/qs/tiles/HotspotTile.java