Add check whether wallpaper's size has changed
If we are in an activity that shows wallpaper(such as Launcher),
when rotating screen, there will be a short time that wallpaper
shows improperly, maybe showing black in some places.

When the wallpaper's size has changed, wallpaper will redraw itself,
which is not contained in the original rotating logic.

To avoid this, we add check when rotating screen, if wallpaper needs to
be resized, wait for wallpaper redraw itself, then stop freezing screen.

Google issue:
https://code.google.com/p/android/issues/detail?id=205219

Change-Id: I12cea9df9ab7b78799f9c080d5adef09ca341ad6
Signed-off-by: placiano <placiano80@gmail.com>