Framework: Fix for file delete on settings not reflecting in MTP host
When file on external sdcard is removed from Settings->
Storage->External sdcard , with the device connected to host
machine in MTP mode. The delete file is still seen on the
MTP host. The issue happens because the external sdcard file
entry is not removed from MediaProvider DB files table. The
failure to remove is due to path mismatch. Fix provided
ensures visible path is used to remove the document.

Change-Id: I526eb4dbbfda79a0641c516e347305c760602efb
Signed-off-by: Shasidhar Ganiga <sganig@codeaurora.org>