revert: revert: Alarm: Add one more RTC alarm type for poweroff alarm.
Now we use type of RTC_WAKEUP for the poweroff alarm.
Actually lots of other applications also use this type.
But these applications never expect the device will
power on after it has been shut down. So we need to
add dedicated alarm type for the Deskclock or similar
applications.

Change-Id: Ib515d8a737ab73cdd809597ad8cde57030c0a0a1 (reverted from commit ed29480f3ce6176ee363cedcb8f4ae2fa51f3e5d) (reverted from commit 8f9b83c677363bc73207505b26e4bb26755d9e3f)