Fix keyguard prompt reason showing with fingerprint reboot enabled
When this option is enabled, the user should not see any message
about a PIN being required on reboot or timeout.

Also, clean up the method of the original patch by moving the
function to the first part of the return statement, as Java will
evaluate it first and move on if when true since it's an OR gate.

Change-Id: I3d498ecf3a4db2abfeb6c21797829f2b57758736
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
(cherry picked from commit 73d5dc445cb5741791c978dc9b997815c875ae43)